[FIX] hr_timesheet, timer: Fix traceback in kanban view when using the timer
Currently, when the unit amount of line is zero and the timer time
spent on entry is less than one minutes then entries are unlinked
once the timer is stopped and due to that the entry disappear from the
kanban view and traceback is generated.

After this commit, we fix the issue by preventing the deletion of
entries by setting options 'prevent_deletion' on widget so we should
no allow to delete the entry if the option is set on widget.

closes odoo/odoo#50395

Taskid: 2230155
Closes: #49128
X-original-commit: 458c39d84c065e67d791ee3ede5279c3735c65d7
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>