[FIX] website: fix go to website on create
When creating a record and clicking on the go to website button before
saving it previously saved the record and browsed it on the website.
Currently it does not save it and the redirection is broken too.

This is fixed by using the same trigger as the other stat buttons.

Introduced by: https://github.com/odoo/odoo/pull/34261

closes odoo/odoo#49262

X-original-commit: 700f57ef13e45527c2d249fc431e1d54e0bc1955
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>