[FIX] survey: block selection by key when already answered in session
Before this commit, during a session, it was still possible to select another
choice (in a multiple choice question) even after having selected and submitted
an answer.

After this commit, if the question has already been answered or time limit
already ran out, it's not possible to use key to select another choice.

Task ID: 2230980

X-original-commit: 3a80d3fd44c3f182204eba98f12bb6ebcf0da59a